# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: PlayZone Bar API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /gallery:
    get:
      operationId: listGallery
      tags: [gallery]
      summary: List gallery images
      parameters:
        - name: activeOnly
          in: query
          schema:
            type: boolean
      responses:
        "200":
          description: Gallery images
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/GalleryImage"
    post:
      operationId: createGalleryImage
      tags: [gallery]
      summary: Add a gallery image
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/GalleryImageInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GalleryImage"
  /gallery/{id}:
    patch:
      operationId: updateGalleryImage
      tags: [gallery]
      summary: Update a gallery image
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/GalleryImageUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GalleryImage"
    delete:
      operationId: deleteGalleryImage
      tags: [gallery]
      summary: Delete a gallery image
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  # MENU CATEGORIES
  /menu/categories:
    get:
      operationId: listMenuCategories
      tags: [menu]
      summary: List menu categories
      responses:
        "200":
          description: Categories
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/MenuCategory"
    post:
      operationId: createMenuCategory
      tags: [menu]
      summary: Create menu category
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/MenuCategoryInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MenuCategory"
  /menu/categories/{id}:
    patch:
      operationId: updateMenuCategory
      tags: [menu]
      summary: Update menu category
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/MenuCategoryUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MenuCategory"
    delete:
      operationId: deleteMenuCategory
      tags: [menu]
      summary: Delete menu category
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  # MENU ITEMS
  /menu/items:
    get:
      operationId: listMenuItems
      tags: [menu]
      summary: List menu items
      parameters:
        - name: categoryId
          in: query
          schema:
            type: integer
        - name: availableOnly
          in: query
          schema:
            type: boolean
      responses:
        "200":
          description: Menu items
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/MenuItem"
    post:
      operationId: createMenuItem
      tags: [menu]
      summary: Create menu item
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/MenuItemInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MenuItem"
  /menu/items/{id}:
    patch:
      operationId: updateMenuItem
      tags: [menu]
      summary: Update menu item
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/MenuItemUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MenuItem"
    delete:
      operationId: deleteMenuItem
      tags: [menu]
      summary: Delete menu item
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  # EVENTS
...
components:
  schemas:
...
    GalleryImage:
      type: object
      required: [id, url, publicId, active, order, createdAt]
      properties:
        id:
          type: integer
        url:
          type: string
        publicId:
          type: string
        active:
          type: boolean
        order:
          type: integer
        alt:
          type: ["string", "null"]
        createdAt:
          type: string
    GalleryImageInput:
      type: object
      required: [url, publicId]
      properties:
        url:
          type: string
        publicId:
          type: string
        active:
          type: boolean
        order:
          type: integer
        alt:
          type: string
    GalleryImageUpdate:
      type: object
      properties:
        active:
          type: boolean
        order:
          type: integer
        alt:
          type: string
    MenuCategory:
      type: object
      required: [id, name, order]
      properties:
        id:
          type: integer
        name:
          type: string
        order:
          type: integer
        icon:
          type: ["string", "null"]
    MenuCategoryInput:
      type: object
      required: [name]
      properties:
        name:
          type: string
        order:
          type: integer
        icon:
          type: string
    MenuCategoryUpdate:
      type: object
      properties:
        name:
          type: string
        order:
          type: integer
        icon:
          type: string
    MenuItem:
      type: object
      required: [id, categoryId, name, price, available]
      properties:
        id:
          type: integer
        categoryId:
          type: integer
        categoryName:
          type: ["string", "null"]
        name:
          type: string
        description:
          type: ["string", "null"]
        price:
          type: number
        imageUrl:
          type: ["string", "null"]
        available:
          type: boolean
    MenuItemInput:
      type: object
      required: [categoryId, name, price]
      properties:
        categoryId:
          type: integer
        name:
          type: string
        description:
          type: string
        price:
          type: number
        imageUrl:
          type: string
        available:
          type: boolean
    MenuItemUpdate:
      type: object
      properties:
        categoryId:
          type: integer
        name:
          type: string
        description:
          type: string
        price:
          type: number
        imageUrl:
          type: string
        available:
          type: boolean
...