# lib/api-zod/src/generated/types/galleryImageUpdate.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * PlayZone Bar API
 * OpenAPI spec version: 0.1.0
 */

export interface GalleryImageUpdate {
  active?: boolean;
  order?: number;
  alt?: string;
}